update smart contract files and hooks
add contract data
