+++ client/package.json
@@ -17,8 +17,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.21.3",
-    "react-toastify": "^10.0.3",
     "react-stately": "^3.29.1",
+    "react-toastify": "^10.0.3",
     "viem": "1.x",
     "wagmi": "1.x"
   },
