+++ client/src/.env
@@ -4,4 +4,4 @@ VITE_INFURA_ID="5db5a885bf0f4fce97af4d59f6bd0e11"
 VITE_GHO_TOKEN_SEPOLIA="0xc4bF5CbDaBE595361438F8c6a187bDc330539c60"
 VITE_XOC_TOKEN_SEPOLIA="insert_address"
 
-VITE_BOOMIE_RAMP_CONTRACT_SEPOLIA="0xe1F5172C6Fca31047411B75C3150F47ad1A80594"
+VITE_BOOMIE_RAMP_CONTRACT_SEPOLIA="0xd73ee34ebaeC74e1f86546D12FB597F962F44680"
