+++ client/src/App.tsx
@@ -1,12 +1,12 @@
-import { WagmiConfig, createConfig } from 'wagmi'
-import { ConnectKitProvider, getDefaultConfig } from 'connectkit'
-import { mainnet, goerli, sepolia } from 'wagmi/chains'
-import { RouterProvider, createBrowserRouter } from 'react-router-dom'
-import { Home, MakeIntent, UserPanel } from './views'
-import { ToastContainer } from 'react-toastify'
-import { Layout } from './components'
-import { MakeDeposit } from './views/MakeDeposit'
-const chains = [mainnet, goerli, sepolia]
+import { WagmiConfig, createConfig } from 'wagmi';
+import { ConnectKitProvider, getDefaultConfig } from 'connectkit';
+import { sepolia } from 'wagmi/chains';
+import { RouterProvider, createBrowserRouter } from 'react-router-dom';
+import { Home, MakeIntent, UserPanel } from './views';
+import { ToastContainer } from 'react-toastify';
+import { Layout } from './components';
+import { MakeDeposit } from './views/MakeDeposit';
+const chains = [sepolia];
 const config = createConfig(
   getDefaultConfig({
     infuraId: import.meta.env.VITE_ALCHEMY_ID,
@@ -18,11 +18,11 @@ const config = createConfig(
     appDescription: 'Your App Description',
     appUrl: 'https://family.co', // your app's url
     appIcon: 'https://family.co/logo.png', // your app's icon, no bigger than 1024x1024px (max. 1MB)
-    chains
+    chains,
   })
-)
+);
 
-function App () {
+function App() {
   const router = createBrowserRouter([
     {
       path: '/',
@@ -30,23 +30,23 @@ function App () {
       children: [
         {
           path: '/',
-          element: <Home />
+          element: <Home />,
         },
         {
           path: '/make-intent/:depositId',
-          element: <MakeIntent />
+          element: <MakeIntent />,
         },
         {
           path: '/make-deposit',
-          element: <MakeDeposit />
+          element: <MakeDeposit />,
         },
         {
           path: '/user-panel',
-          element: <UserPanel />
-        }
-      ]
-    }
-  ])
+          element: <UserPanel />,
+        },
+      ],
+    },
+  ]);
 
   return (
     <WagmiConfig config={config}>
@@ -55,7 +55,7 @@ function App () {
         <ToastContainer />
       </ConnectKitProvider>
     </WagmiConfig>
-  )
+  );
 }
 
-export default App
+export default App;
