+++ client/src/lib/contracts/BoomieRampABI.ts
@@ -4,46 +4,46 @@ export const BoomieRampABI = [
       {
         internalType: 'address',
         name: '_owner',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'contract IERC20',
         name: '_gho',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'uint256',
         name: '_minDepositAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: '_maxOnRampAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: '_intentExpirationPeriod',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: '_onRampCooldownPeriod',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: '_sustainabilityFee',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'address',
         name: '_sustainabilityFeeRecipient',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     stateMutability: 'nonpayable',
-    type: 'constructor'
+    type: 'constructor',
   },
   {
     anonymous: false,
@@ -52,17 +52,17 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'address',
         name: 'accountOwner',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: true,
         internalType: 'bytes32',
         name: 'venmoIdHash',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'AccountRegistered',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -71,17 +71,17 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'uint256',
         name: 'depositId',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: false,
         internalType: 'address',
         name: 'sellerAccount',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'DepositClosed',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -90,29 +90,29 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'uint256',
         name: 'depositId',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'sellerAccount',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'amount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'conversionRate',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'DepositReceived',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -121,23 +121,23 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'uint256',
         name: 'depositId',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'sellerAccount',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'amount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'DepositWithdrawn',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -146,11 +146,11 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'uint256',
         name: 'intentExpirationPeriod',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'IntentExpirationPeriodSet',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -159,41 +159,41 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'bytes32',
         name: 'intentHash',
-        type: 'bytes32'
+        type: 'bytes32',
       },
       {
         indexed: true,
         internalType: 'uint256',
         name: 'depositId',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'buyerAccount',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: false,
         internalType: 'address',
         name: 'to',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'amount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'feeAmount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'IntentFulfilled',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -202,17 +202,17 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'bytes32',
         name: 'intentHash',
-        type: 'bytes32'
+        type: 'bytes32',
       },
       {
         indexed: true,
         internalType: 'uint256',
         name: 'depositId',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'IntentPruned',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -221,41 +221,41 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'bytes32',
         name: 'intentHash',
-        type: 'bytes32'
+        type: 'bytes32',
       },
       {
         indexed: true,
         internalType: 'uint256',
         name: 'depositId',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'buyerAccount',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: false,
         internalType: 'address',
         name: 'to',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'amount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         indexed: false,
         internalType: 'uint256',
         name: 'timestamp',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'IntentSignaled',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -264,11 +264,11 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'uint256',
         name: 'maxOnRampAmount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'MaxOnRampAmountSet',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -277,11 +277,11 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'uint256',
         name: 'minDepositAmount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'MinDepositAmountSet',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -290,11 +290,11 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'uint256',
         name: 'onRampCooldownPeriod',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'OnRampCooldownPeriodSet',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -303,17 +303,17 @@ export const BoomieRampABI = [
         indexed: true,
         internalType: 'address',
         name: 'previousOwner',
-        type: 'address'
+        type: 'address',
       },
       {
         indexed: true,
         internalType: 'address',
         name: 'newOwner',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'OwnershipTransferred',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -322,11 +322,11 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'address',
         name: 'feeRecipient',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'SustainabilityFeeRecipientUpdated',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -335,11 +335,11 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'uint256',
         name: 'fee',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'SustainabilityFeeUpdated',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -348,17 +348,17 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'bytes32',
         name: 'listOwner',
-        type: 'bytes32'
+        type: 'bytes32',
       },
       {
         indexed: false,
         internalType: 'bytes32',
         name: 'deniedUser',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'UserAddedToDenylist',
-    type: 'event'
+    type: 'event',
   },
   {
     anonymous: false,
@@ -367,82 +367,87 @@ export const BoomieRampABI = [
         indexed: false,
         internalType: 'bytes32',
         name: 'listOwner',
-        type: 'bytes32'
+        type: 'bytes32',
       },
       {
         indexed: false,
         internalType: 'bytes32',
         name: 'approvedUser',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'UserRemovedFromDenylist',
-    type: 'event'
+    type: 'event',
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_deniedUser',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'addAccountToDenylist',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'allDeposits',
     outputs: [
       {
+        internalType: 'uint256',
+        name: 'depositId',
+        type: 'uint256',
+      },
+      {
         internalType: 'address',
         name: 'depositor',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'uint256',
         name: 'depositAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'remainingDeposits',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'outstandingIntentAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'conversionRate',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_intentHash',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'cancelIntent',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -451,58 +456,63 @@ export const BoomieRampABI = [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'deposits',
     outputs: [
       {
+        internalType: 'uint256',
+        name: 'depositId',
+        type: 'uint256',
+      },
+      {
         internalType: 'address',
         name: 'depositor',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'uint256',
         name: 'depositAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'remainingDeposits',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'outstandingIntentAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'conversionRate',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'getAccountDeposits',
     outputs: [
@@ -511,71 +521,76 @@ export const BoomieRampABI = [
           {
             internalType: 'uint256',
             name: 'depositId',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'bytes32',
             name: 'depositorIdHash',
-            type: 'bytes32'
+            type: 'bytes32',
           },
           {
             components: [
               {
+                internalType: 'uint256',
+                name: 'depositId',
+                type: 'uint256',
+              },
+              {
                 internalType: 'address',
                 name: 'depositor',
-                type: 'address'
+                type: 'address',
               },
               {
                 internalType: 'uint256',
                 name: 'depositAmount',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'remainingDeposits',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'outstandingIntentAmount',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'conversionRate',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'bytes32[]',
                 name: 'intentHashes',
-                type: 'bytes32[]'
-              }
+                type: 'bytes32[]',
+              },
             ],
             internalType: 'struct BoomieRamp.Deposit',
             name: 'deposit',
-            type: 'tuple'
+            type: 'tuple',
           },
           {
             internalType: 'uint256',
             name: 'availableLiquidity',
-            type: 'uint256'
-          }
+            type: 'uint256',
+          },
         ],
         internalType: 'struct BoomieRamp.DepositWithAvailableLiquidity[]',
         name: 'accountDeposits',
-        type: 'tuple[]'
-      }
+        type: 'tuple[]',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'getAccountInfo',
     outputs: [
@@ -584,118 +599,123 @@ export const BoomieRampABI = [
           {
             internalType: 'bytes32',
             name: 'venmoIdHash',
-            type: 'bytes32'
+            type: 'bytes32',
           },
           {
             internalType: 'uint256[]',
             name: 'deposits',
-            type: 'uint256[]'
-          }
+            type: 'uint256[]',
+          },
         ],
         internalType: 'struct BoomieRamp.AccountInfo',
         name: '',
-        type: 'tuple'
-      }
+        type: 'tuple',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'getAccountVenmoIdHash',
     outputs: [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'getDeniedUsers',
     outputs: [
       {
         internalType: 'bytes32[]',
         name: '',
-        type: 'bytes32[]'
-      }
+        type: 'bytes32[]',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_depositId',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'getDeposit',
     outputs: [
       {
         components: [
           {
+            internalType: 'uint256',
+            name: 'depositId',
+            type: 'uint256',
+          },
+          {
             internalType: 'address',
             name: 'depositor',
-            type: 'address'
+            type: 'address',
           },
           {
             internalType: 'uint256',
             name: 'depositAmount',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'uint256',
             name: 'remainingDeposits',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'uint256',
             name: 'outstandingIntentAmount',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'uint256',
             name: 'conversionRate',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'bytes32[]',
             name: 'intentHashes',
-            type: 'bytes32[]'
-          }
+            type: 'bytes32[]',
+          },
         ],
         internalType: 'struct BoomieRamp.Deposit',
         name: '',
-        type: 'tuple'
-      }
+        type: 'tuple',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256[]',
         name: '_depositIds',
-        type: 'uint256[]'
-      }
+        type: 'uint256[]',
+      },
     ],
     name: 'getDepositFromIds',
     outputs: [
@@ -704,63 +724,68 @@ export const BoomieRampABI = [
           {
             internalType: 'uint256',
             name: 'depositId',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'bytes32',
             name: 'depositorIdHash',
-            type: 'bytes32'
+            type: 'bytes32',
           },
           {
             components: [
               {
+                internalType: 'uint256',
+                name: 'depositId',
+                type: 'uint256',
+              },
+              {
                 internalType: 'address',
                 name: 'depositor',
-                type: 'address'
+                type: 'address',
               },
               {
                 internalType: 'uint256',
                 name: 'depositAmount',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'remainingDeposits',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'outstandingIntentAmount',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'conversionRate',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'bytes32[]',
                 name: 'intentHashes',
-                type: 'bytes32[]'
-              }
+                type: 'bytes32[]',
+              },
             ],
             internalType: 'struct BoomieRamp.Deposit',
             name: 'deposit',
-            type: 'tuple'
+            type: 'tuple',
           },
           {
             internalType: 'uint256',
             name: 'availableLiquidity',
-            type: 'uint256'
-          }
+            type: 'uint256',
+          },
         ],
         internalType: 'struct BoomieRamp.DepositWithAvailableLiquidity[]',
         name: 'depositArray',
-        type: 'tuple[]'
-      }
+        type: 'tuple[]',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -769,51 +794,56 @@ export const BoomieRampABI = [
       {
         components: [
           {
+            internalType: 'uint256',
+            name: 'depositId',
+            type: 'uint256',
+          },
+          {
             internalType: 'address',
             name: 'depositor',
-            type: 'address'
+            type: 'address',
           },
           {
             internalType: 'uint256',
             name: 'depositAmount',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'uint256',
             name: 'remainingDeposits',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'uint256',
             name: 'outstandingIntentAmount',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'uint256',
             name: 'conversionRate',
-            type: 'uint256'
+            type: 'uint256',
           },
           {
             internalType: 'bytes32[]',
             name: 'intentHashes',
-            type: 'bytes32[]'
-          }
+            type: 'bytes32[]',
+          },
         ],
         internalType: 'struct BoomieRamp.Deposit[]',
         name: '',
-        type: 'tuple[]'
-      }
+        type: 'tuple[]',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'bytes32[]',
         name: '_intentHashes',
-        type: 'bytes32[]'
-      }
+        type: 'bytes32[]',
+      },
     ],
     name: 'getIntentsWithOnRamperId',
     outputs: [
@@ -822,91 +852,91 @@ export const BoomieRampABI = [
           {
             internalType: 'bytes32',
             name: 'intentHash',
-            type: 'bytes32'
+            type: 'bytes32',
           },
           {
             components: [
               {
                 internalType: 'address',
                 name: 'buyerAccount',
-                type: 'address'
+                type: 'address',
               },
               {
                 internalType: 'address',
                 name: 'to',
-                type: 'address'
+                type: 'address',
               },
               {
                 internalType: 'uint256',
                 name: 'deposit',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'amount',
-                type: 'uint256'
+                type: 'uint256',
               },
               {
                 internalType: 'uint256',
                 name: 'intentTimestamp',
-                type: 'uint256'
-              }
+                type: 'uint256',
+              },
             ],
             internalType: 'struct BoomieRamp.Intent',
             name: 'intent',
-            type: 'tuple'
+            type: 'tuple',
           },
           {
             internalType: 'bytes32',
             name: 'onRamperIdHash',
-            type: 'bytes32'
-          }
+            type: 'bytes32',
+          },
         ],
         internalType: 'struct BoomieRamp.IntentWithOnRamperId[]',
         name: '',
-        type: 'tuple[]'
-      }
+        type: 'tuple[]',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'getLastOnRampTimestamp',
     outputs: [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'getVenmoIdCurrentIntentHash',
     outputs: [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -915,18 +945,18 @@ export const BoomieRampABI = [
       {
         internalType: 'contract IERC20',
         name: '',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
     name: 'initialize',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -935,74 +965,74 @@ export const BoomieRampABI = [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'intents',
     outputs: [
       {
         internalType: 'address',
         name: 'buyerAccount',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'address',
         name: 'to',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'uint256',
         name: 'deposit',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'amount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: 'intentTimestamp',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_account',
-        type: 'address'
+        type: 'address',
       },
       {
         internalType: 'bytes32',
         name: '_deniedUser',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'isDeniedUser',
     outputs: [
       {
         internalType: 'bool',
         name: '',
-        type: 'bool'
-      }
+        type: 'bool',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1011,11 +1041,11 @@ export const BoomieRampABI = [
       {
         internalType: 'bool',
         name: '',
-        type: 'bool'
-      }
+        type: 'bool',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1024,11 +1054,11 @@ export const BoomieRampABI = [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1037,42 +1067,42 @@ export const BoomieRampABI = [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_depositAmount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: '_receiveAmount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'offRamp',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_intentHash',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'onRamp',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1081,11 +1111,11 @@ export const BoomieRampABI = [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1094,152 +1124,152 @@ export const BoomieRampABI = [
       {
         internalType: 'address',
         name: '',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
     name: 'register',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_intentHash',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'releaseFundsToOnramper',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'bytes32',
         name: '_approvedUser',
-        type: 'bytes32'
-      }
+        type: 'bytes32',
+      },
     ],
     name: 'removeAccountFromDenylist',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
     name: 'renounceOwnership',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_intentExpirationPeriod',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'setIntentExpirationPeriod',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_maxOnRampAmount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'setMaxOnRampAmount',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_minDepositAmount',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'setMinDepositAmount',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_onRampCooldownPeriod',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'setOnRampCooldownPeriod',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_fee',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     name: 'setSustainabilityFee',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: '_feeRecipient',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'setSustainabilityFeeRecipient',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256',
         name: '_depositId',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'uint256',
         name: '_amount',
-        type: 'uint256'
+        type: 'uint256',
       },
       {
         internalType: 'address',
         name: '_to',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'signalIntent',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1248,11 +1278,11 @@ export const BoomieRampABI = [
       {
         internalType: 'uint256',
         name: '',
-        type: 'uint256'
-      }
+        type: 'uint256',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [],
@@ -1261,40 +1291,40 @@ export const BoomieRampABI = [
       {
         internalType: 'address',
         name: '',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     stateMutability: 'view',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'address',
         name: 'newOwner',
-        type: 'address'
-      }
+        type: 'address',
+      },
     ],
     name: 'transferOwnership',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     inputs: [
       {
         internalType: 'uint256[]',
         name: '_depositIds',
-        type: 'uint256[]'
-      }
+        type: 'uint256[]',
+      },
     ],
     name: 'withdrawDeposit',
     outputs: [],
     stateMutability: 'nonpayable',
-    type: 'function'
+    type: 'function',
   },
   {
     stateMutability: 'payable',
-    type: 'receive'
-  }
-]
+    type: 'receive',
+  },
+];
