+++ client/src/main.tsx
@@ -1,9 +1,8 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
-import './index.css'
-import { NextUIProvider } from '@nextui-org/react'
-
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App.tsx';
+import './index.css';
+import { NextUIProvider } from '@nextui-org/react';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
@@ -11,4 +10,4 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
       <App />
     </NextUIProvider>
   </React.StrictMode>
-)
+);
