+++ client/src/views/Home.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React from 'react';
 import {
   Table,
   TableHeader,
@@ -9,9 +9,21 @@ import {
   Pagination,
   Spinner,
   getKeyValue,
-  Button
-} from '@nextui-org/react'
-import { useNavigate } from 'react-router-dom'
+  Button,
+} from '@nextui-org/react';
+import { useNavigate } from 'react-router-dom';
+import { useGetDepositsWithMinimumBalance } from '@/hooks';
+import { formatEther } from 'viem';
+
+type Deposit = {
+  depositId: bigint;
+  depositor: `0x${string}`;
+  depositAmount: bigint;
+  remainingDeposits: bigint;
+  outstandingIntentAmount: bigint;
+  conversionRate: bigint;
+  intentHashes: `0x${string}`[];
+};
 
 const mockData = [
   {
@@ -20,7 +32,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 2,
@@ -28,7 +40,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 3,
@@ -36,7 +48,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 4,
@@ -44,7 +56,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 5,
@@ -52,7 +64,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 6,
@@ -60,7 +72,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 7,
@@ -68,7 +80,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 8,
@@ -76,7 +88,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 9,
@@ -84,7 +96,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 10,
@@ -92,7 +104,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 11,
@@ -100,7 +112,7 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
+    outstandingIntentAmount: 70,
   },
   {
     id: 12,
@@ -108,44 +120,62 @@ const mockData = [
     depositAmount: 100,
     remainingDeposits: 78,
     conversionRate: 17.75,
-    outstandingIntentAmount: 70
-  }
-]
+    outstandingIntentAmount: 70,
+  },
+];
 
-function Home () {
-  const [page, setPage] = React.useState(1)
-  const rowsPerPage = 10
-  const navigate = useNavigate()
+function Home() {
+  const [page, setPage] = React.useState(1);
+  const rowsPerPage = 10;
+  const navigate = useNavigate();
 
-  const pages = Math.ceil(mockData.length / rowsPerPage)
+  const pages = Math.ceil(mockData.length / rowsPerPage);
+
+  const { getDepositsWithMinimumBalanceData } =
+    useGetDepositsWithMinimumBalance();
 
   const items = React.useMemo(() => {
-    const start = (page - 1) * rowsPerPage
-    const end = start + rowsPerPage
+    const start = (page - 1) * rowsPerPage;
+    const end = start + rowsPerPage;
+    if (!getDepositsWithMinimumBalanceData) return mockData.slice(start, end);
+    console.log(getDepositsWithMinimumBalanceData as Deposit[]);
+    return (getDepositsWithMinimumBalanceData as Deposit[])
+      .map((deposit) => {
+        return {
+          id: formatEther(deposit.depositId),
+          depositor: deposit.depositor,
+          depositAmount: formatEther(deposit.depositAmount),
+          remainingDeposits: formatEther(deposit.remainingDeposits),
+          outstandingIntentAmount: formatEther(deposit.outstandingIntentAmount),
+          conversionRate: formatEther(deposit.conversionRate),
+          intentHashes: deposit.intentHashes,
+        };
+      })
+      .slice(start, end);
+  }, [page, getDepositsWithMinimumBalanceData]);
 
-    return mockData.slice(start, end)
-  }, [page, mockData])
+  // console.log(getDepositsWithMinimumBalanceData);
 
-  const handleOnClickButton = (id: number) => {
-    navigate(`/make-intent/${id}`)
-  }
+  const handleOnClickButton = (id: string) => {
+    navigate(`/make-intent/${id}`);
+  };
 
   return (
-    <main className='flex flex-col justify-center min-h-screen gap-5 items-center'>
+    <main className="flex flex-col justify-center min-h-screen gap-5 items-center">
       <Table
-        className='w-[80vw]'
+        className="w-[80vw]"
         isHeaderSticky
-        aria-label='Órdenes top'
+        aria-label="Órdenes top"
         bottomContent={
-          <div className='flex w-full justify-center'>
+          <div className="flex w-full justify-center">
             <Pagination
               isCompact
               showControls
               showShadow
-              color='secondary'
+              color="secondary"
               page={page}
               total={pages}
-              onChange={page => setPage(page)}
+              onChange={(page) => setPage(page)}
             />
           </div>
         }
@@ -153,50 +183,50 @@ function Home () {
           base: 'max-h-[820px] overflow-scroll',
           table: 'max-h-[60vh]',
           th: 'bg-secondary text-white font-bold',
-          wrapper: ' bg-card-bg'
+          wrapper: ' bg-card-bg',
         }}
       >
         <TableHeader>
-          <TableColumn key='conversionRate'>PRECIO</TableColumn>
-          <TableColumn key='remainingDeposits'>DISPONIBLE</TableColumn>
-          <TableColumn key='depositor'>VENDEDOR</TableColumn>
-          <TableColumn key='outstandingIntentAmount'>
+          <TableColumn key="conversionRate">PRECIO</TableColumn>
+          <TableColumn key="remainingDeposits">DISPONIBLE</TableColumn>
+          <TableColumn key="depositor">VENDEDOR</TableColumn>
+          <TableColumn key="outstandingIntentAmount">
             ÓRDENES EXITOSAS
           </TableColumn>
-          <TableColumn key='action'>ACCIÓN</TableColumn>
+          <TableColumn key="action">ACCIÓN</TableColumn>
         </TableHeader>
         <TableBody items={items} loadingContent={<Spinner />}>
-          {item => (
+          {(item) => (
             <TableRow key={item.id} accessKey={item.id.toString()}>
-              {columnKey => {
+              {(columnKey) => {
                 if (columnKey === 'action') {
                   return (
                     <TableCell>
                       <Button
                         onClick={() => handleOnClickButton(item.id)}
-                        className='w-full bg-main text-white font-bold'
+                        className="w-full bg-main text-white font-bold"
                       >
                         Comprar
                       </Button>
                     </TableCell>
-                  )
+                  );
                 }
                 if (columnKey === 'conversionRate') {
-                  return <TableCell>${item[columnKey]} MXN</TableCell>
+                  return <TableCell>${item[columnKey]} MXN</TableCell>;
                 }
 
                 if (columnKey === 'remainingDeposits') {
-                  return <TableCell>{item[columnKey]} GHO</TableCell>
+                  return <TableCell>{item[columnKey]} GHO</TableCell>;
                 }
 
-                return <TableCell>{getKeyValue(item, columnKey)}</TableCell>
+                return <TableCell>{getKeyValue(item, columnKey)}</TableCell>;
               }}
             </TableRow>
           )}
         </TableBody>
       </Table>
     </main>
-  )
+  );
 }
 
-export { Home }
+export { Home };
